feat(compose): getViewStrategy hook added
This commit takes advantage of new capabilities of the resource
coordinator which allow loading a view model type independently of
anonymous element creation. This allows manual instantiation of the
view model. Once this is accomplished it is possible to call a
getViewStrategy hook to allow per-view-model data-based determination
of views.