fix(GlobalBehavior): throw AggregateError
fixed incorrect passing of inner error to Error constructor

note: at some point the AggregateError function should be renamed to aggregateError, and all calls updated also.  AggregateError is a function that returns an aggregate error, not a class.  a common mistake associated with the incorrect name is the incorrect use of new when calling the function