feature(dirty-binding-behavior): Add a dirty binding behavior that regularly checks for updates
When binding arbitrary expressions that depend on complex objects there are no tools to either observe the dependencies or force a regular update. The dirty binding behavior adds a binding behavior that overrides the default binding behavior and instead forces regular updates. The default update interval is 100ms, which research has shown is the threshold for an action to be perceived as instantaneous (see here: https://ux.stackexchange.com/questions/2/what-is-an-acceptable-response-time-for-my-ajax-ui/4#4). Developers can customize the responsiveness of the binding behavior by specifying a different update interval in ms as a parameter of the binding behavior.

Examples:

100ms, or 10fps: <div>${obj | toJson & dirty}</div>
16ms, or 60fps: <div>{${obj | toJson & dirty: 16}</div>